# control/swagger.yaml
basePath: /
consumes:
    - application/json
securityDefinitions:
    cookieAuth:
        type: apiKey
        in: cookie
        name: Access-Token
    headerAuth:
        type: apiKey
        in: header
        name: X-Access-Token
security:
    - cookieAuth: []
    - headerAuth: []
definitions:
...
    Profits:
        type: object
        description: Profits contains the profit breakdown for a given portfolio
        properties:
            Name: 
                type: string
                example: "401(k)" 
            OriginalValue:
                type: float
                example: 218.4
            NewValue:
                type: float
                example: 37047.6
            NetGain:
                type: float
                example: 35651.43
            NetChange:
                type: string
                example: "1221%"
            Moves:
                type: array
                items:
                    '#/definitions/Security'
                example:
                    - Ticker: "VOO"
                      BoughtPrice: 10.92
                      CurrPrice: 308.73
                      Shares: 100
                      Gain: 29700.81
                      Change: "2827%"
                      Currency: "USD"
                    - Ticker: "AAPL"
                      BoughtPrice: 10.92
                      CurrPrice: 308.73
                      Shares: 20
                      Gain: 5950.62
                      Change: "2827%"
                      Currency: "USD"
             
                

        x-go-package: github.com/fercevik729/STLKER/control/handlers
    Security:
        type: object
        description: Security represents a single security and all of its financial details
        properties:
            Ticker:
                type: string
                example: "TSLA"
            BoughtPrice: 
                type: float
                example: 10.92 
            CurrPrice: 
                type: string
                example: 308.73
            Shares:
                type: float
                example: 1 
            Gain: 
                type: float
                example: 297.81
            Change: 
                type: string
                example: "2827%"
            Currency: 
                type: string
                example: "USD"

        x-go-package: github.com/fercevik729/STLKER/control/handlers
    ReqSecurity:
        type: object
        description: ReqSecurity defines the schema for a Security in a request payload
        required:
            - ticker
            - shares
        properties:
            ticker:
                type: string
                example: "F"
            shares:
                type: float
                example: 5.75
...
info:
    description: Documentation for STLKER API
    title: of STLKER API
    version: 1.0.0
paths:
...
    /portfolios/{name}:
        delete:
            description: Deletes a given portfolio for a user
            operationId: deletePortfolio
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "200":
                    $ref: '#/responses/messageResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - portfolios
            security:
                - cookieAuth: []
                - headerAuth: []
        get:
            description: Outputs a particular portfolio for a user
            operationId: getPortfolio
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "200":
                    $ref: '#/responses/profitResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - portfolios
            security:
                - cookieAuth: []
                - headerAuth: []
        post:
            description: Creates a new security
            operationId: createSecurity
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  example: "CollegeFund"
                  x-go-name: Name
                - description: The new security
                  in: body
                  required: true
                  name: Body
                  schema:
                    $ref: '#/definitions/ReqSecurity'
            responses:
                "201":
                    $ref: '#/responses/messageResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - securities
            security:
                - cookieAuth: []
                - headerAuth: []
        put:
            description: Updates a security's information for a given portfolio
            operationId: updateSecurity
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
                - description: The updated security
                  in: body
                  required: true
                  name: Body
                  schema:
                    $ref: '#/definitions/ReqSecurity'
            responses:
                "201":
                    $ref: '#/responses/messageResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - securities
            security:
                - cookieAuth: []
                - headerAuth: []
    /portfolios/{name}/{ticker}:
        delete:
            description: Deletes a security from a given portfolio
            operationId: deleteSecurity
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
                - description: Ticker of the security
                  in: path
                  name: ticker
                  required: true
                  type: string
                  x-go-name: Ticker
            responses:
                "201":
                    $ref: '#/responses/messageResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - securities
            security:
                - cookieAuth: []
                - headerAuth: []
        get:
            description: Outputs a security's details to the client
            operationId: readSecurity
            parameters:
                - description: Name of the portfolio
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
                - description: Ticker of the security
                  in: path
                  name: ticker
                  required: true
                  type: string
                  x-go-name: Ticker
            responses:
                "200":
                    $ref: '#/responses/securityResponse'
                "400":
                    $ref: '#/responses/clientErrResponse'
                "401":
                    $ref: '#/responses/unAuthResponse'
                "500":
                    $ref: '#/responses/serverErrResponse'
            tags:
                - securities
            security:
                - cookieAuth: []
                - headerAuth: []
...
produces:
    - application/json
responses:
...
    clientErrResponse:
        description: A clientErrResponse represents a 4** http error
    serverErrResponse:
        description: An serverErrResponse represents a 5** http error
    unAuthResponse:
        description: An unAuthResponse signifies that the user is unauthorized to make a request
    messageResponse:
        description: A messageResponse returns diagnostic information to the client
...
    profitResponse:
        description: A profitResponse is a single portfolio returned to the client
        schema:
          $ref: '#/definitions/Profits'
...
    securityResponse:
        description: A securityResponse is a single security's information returned to the client
        schema:
          $ref: '#/definitions/Security'
...
schemes:
    - http
swagger: "2.0"